:recycle: Convert users app into purely CMS app
Django CMS no longer fetches the 'parent' page for these pages that are
not in the CMS page tree, breaking the menus/header etc. So far, the
only way seems to make them into an apphook, but that breaks the main
templates if the apphook is not installed. So, be careful.